ml/requirements.txt
# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
albumentations>=1.3.1

# Training Utilities
tqdm>=4.66.0
tensorboard>=2.15.0
wandb>=0.16.0

# Model Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
hydra-core>=1.3.2
omegaconf>=2.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
loguru>=0.7.2
rich>=13.7.0
